+++ lib/src/category/cdc/model/wtstat.dart
@@ -72,7 +72,7 @@ sealed class CDCWeightForStature extends LengthBasedResult
       checkLength(_adjustedLength, contextData: contextData);
 
   num get _length {
-    final val = lengthAtObservationDate.value;
+    final val = lengthAtObservationDate.value.toDouble();
     final gridKey = (val * 2).round() / 2;
     if (contextData._data[sex]?.containsKey(gridKey) ?? false) {
       return gridKey;
@@ -94,7 +94,7 @@ sealed class CDCWeightForStature extends LengthBasedResult
   _CDCWeightForStatureLMS get lmsData => _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => weight.toKilogram.value;
+  num get measurementResultInDefaultUnit => weight.toKilogram.value.toDouble();
 }
 
 class _CDCWeightForStatureLMS extends LMSContext {
