+++ lib/src/category/intergrowth/model/at_birth/newborn_size_for_very_preterm_infants/weight.dart
@@ -88,7 +88,8 @@ sealed class IntergrowthVeryPretermNewbornBirthWeightForAge
   _IntergrowthVeryPretermNewbornBirthWeightForAgeLMS get lmsData => _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => measurementResult.toKilogram.value;
+  num get measurementResultInDefaultUnit =>
+      measurementResult.toKilogram.value.toDouble();
 }
 
 class _IntergrowthVeryPretermNewbornBirthWeightForAgeLMS extends LMSContext {
