+++ lib/src/category/intergrowth/model/during_pregnancy/fetal_growth/abdominal_circumference.dart
@@ -86,7 +86,7 @@ sealed class IntergrowthFetalGrowtAbdominalCircumferenceForAge
 
   @override
   num get measurementResultInDefaultUnit =>
-      measurementResult.toMillimeter.value;
+      measurementResult.toMillimeter.value.toDouble();
 }
 
 class _IntergrowthFetalGrowtAbdominalCircumferenceForAgeLMS extends LMSContext {
