+++ lib/src/category/intergrowth/model/during_pregnancy/fetal_growth/occipitofrontal_diameter.dart
@@ -88,7 +88,7 @@ sealed class IntergrowthFetalGrowtOccipitoFrontalDiameterForAge
 
   @override
   num get measurementResultInDefaultUnit =>
-      measurementResult.toMillimeter.value;
+      measurementResult.toMillimeter.value.toDouble();
 }
 
 class _IntergrowthFetalGrowtOccipitoFrontalDiameterForAgeLMS
