+++ lib/src/category/intergrowth/model/during_pregnancy/fetal_size_in_early_pregnancy/fetal_size_in_early_pregnancy.dart
@@ -74,7 +74,8 @@ sealed class IntergrowthFetalCrownRumpLengthForAge extends AgeBasedResult
   _IntergrowthFetalCrownRumpLengthForAgeLMS get lmsData => _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => length.toMillimeter.value;
+  num get measurementResultInDefaultUnit =>
+      length.toMillimeter.value.toDouble();
 }
 
 class _IntergrowthFetalCrownRumpLengthForAgeLMS extends LMSContext {
