+++ lib/src/category/intergrowth/model/postnatal/postnatal_growth_of_preterm_infants/weight.dart
@@ -98,7 +98,8 @@ sealed class IntergrowthPretermInfantsPostnatalGrowthWeightForAge
       _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => measurementResult.toKilogram.value;
+  num get measurementResultInDefaultUnit =>
+      measurementResult.toKilogram.value.toDouble();
 }
 
 class _IntergrowthPretermInfantsPostnatalGrowthWeightForAgeLMS
