+++ lib/src/category/who/standard/model/body_mass_index_for_age.dart
@@ -120,7 +120,8 @@ sealed class WHOGrowthStandardsBodyMassIndexForAge extends AgeBasedResult
   _WHOGrowthStandardsBodyMassIndexForAgeLMS get lmsData => _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => bodyMassIndexMeasurement.value;
+  num get measurementResultInDefaultUnit =>
+      bodyMassIndexMeasurement.value.toDouble();
 }
 
 class _WHOGrowthStandardsBodyMassIndexForAgeLMS extends LMSContext {
