+++ lib/src/category/who/standard/model/length_for_age.dart
@@ -69,7 +69,7 @@ sealed class WHOGrowthStandardsLengthForAge extends AgeBasedResult
     type: AdjustedLengthType.who,
   );
 
-  num get _zScore => _ageData.lms.zScore(_adjustedLength.value);
+  num get _zScore => _ageData.lms.zScore(_adjustedLength.value.toDouble());
 
   @override
   Age get ageAtObservationDate => checkAge(age, contextData: contextData);
@@ -86,7 +86,7 @@ sealed class WHOGrowthStandardsLengthForAge extends AgeBasedResult
   _WHOGrowthStandardsLengthForAgeLMS get lmsData => _ageData;
 
   @override
-  num get measurementResultInDefaultUnit => _adjustedLength.value;
+  num get measurementResultInDefaultUnit => _adjustedLength.value.toDouble();
 }
 
 class _WHOGrowthStandardsLengthForAgeLMS extends LMSContext {
